Fix `Uncaught TypeError` occurring on Firefox
In `loader.js`, the expression:

> `event.toElement`

causes the following error message on Firefox when the user invokes any
action in the UI:

```
Uncaught TypeError: can't access property "getAttribute", event.toElement is undefined
    downloadTable loader.js:69
    onclick (index):1
```

This is caused by the usage of a browser-specific property, `toElement`,
which appears to exist only on IE and Chrome. [1]

Fix this by using `event.target` instead, the official property that
all browsers recognize. [2]

[1]: https://stackoverflow.com/q/8600174

[2]: https://developer.mozilla.org/en-US/docs/Web/API/Event/target